Restructure Django project
