+++ config/settings/__init__.py
@@ -1,0 +1,1 @@
+# settings package
