+++ pytest.ini
@@ -1,4 +1,4 @@
 [pytest]
-DJANGO_SETTINGS_MODULE = chemically.settings
-pythonpath = chemically
+DJANGO_SETTINGS_MODULE = config.settings.development
+pythonpath = .
 python_files = test_*.py *_test.py
