+++ requirements/base.txt
@@ -1,0 +1,5 @@
+chempy==0.9.0
+Django==5.2.3
+Pint==0.24.4
+pyparsing==3.2.3
+python-dotenv==1.1.0
