+++ requirements/development.txt
@@ -1,0 +1,4 @@
+-r base.txt
+pytest==8.4.0
+pytest-django==4.11.1
+flake8==7.0.0
