+++ static/.gitkeep
@@ -1,0 +1,1 @@
+keep
